Moved _set_db_data_on_span into queries module
